requirements.txt
pandas
spacy
numpy
sentence_transformers
Scikit-learn
pillow
